- Added a post update job to backfill the connected date time for Connection Requests that were marked complete prior to v19.

+++ Rock/Model/Connection/ConnectionType/ConnectionTypeService.cs
@@ -535,7 +535,7 @@ namespace Rock.Model
                     TimelinessPercent = connectedCount == 0
                         ? 0
                         : (
-                            g.Count( x =>
+                            ( decimal ) g.Count( x =>
                                 x.cr.ConnectionState == ConnectionState.Connected
                                 && x.cr.ConnectedDateTime.HasValue
                                 && (
